+++ phpunit-integration-legacy.xml
@@ -74,6 +74,9 @@
         <testsuite name="eZ\Publish\Core\FieldType\Tests\Integration">
             <directory>eZ/Publish/Core/FieldType/Tests/Integration</directory>
         </testsuite>
+        <testsuite name="Ibexa\Tests\Core\Repository\Values\Content">
+            <file>tests/lib/Repository/Values/Content/DefaultLanguageCodeForContentTest.php</file>
+        </testsuite>
         <testsuite name="gateway-integration">
             <file>eZ/Publish/SPI/Tests/Search/Content/IndexerGatewayTest.php</file>
         </testsuite>
